feat: add flat recommended config
recommended.js is a flat config version of the recommended config.

This diverges from the legacy config version of the recommended config
as it does not automatically configure eslint-config-prettier.

Configs extending other configs is complicated and not transparent. We
should trust the user to configure their configs directly. In a future
major change we should remove eslint-config-prettier from the legacy
config too.